refactor: Split aws.go into multiple files
