+++ pkg/clients/config.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2019 The Crossplane Authors.
+Copyright 2023 The Crossplane Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -17,15 +17,10 @@ limitations under the License.
 package aws
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
-	"net"
-	"net/url"
 	"os"
 	"strings"
-	"time"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	awsmiddleware "github.com/aws/aws-sdk-go-v2/aws/middleware"
@@ -34,22 +29,15 @@ import (
 	"github.com/aws/aws-sdk-go-v2/credentials/stscreds"
 	stscredstypesv2 "github.com/aws/aws-sdk-go-v2/service/sts/types"
 
-	ec2type "github.com/aws/aws-sdk-go-v2/service/ec2/types"
 	"github.com/aws/aws-sdk-go-v2/service/sts"
 	awsv1 "github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/awserr"
 	credentialsv1 "github.com/aws/aws-sdk-go/aws/credentials"
 	endpointsv1 "github.com/aws/aws-sdk-go/aws/endpoints"
 	requestv1 "github.com/aws/aws-sdk-go/aws/request"
 	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/aws/smithy-go"
 	"github.com/aws/smithy-go/middleware"
-	jsonpatch "github.com/evanphx/json-patch"
 	"github.com/go-ini/ini"
-	"github.com/google/go-cmp/cmp"
-	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/pkg/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
@@ -790,497 +778,4 @@ func SetWebIdentityRoleOptions(pc *v1beta1.ProviderConfig) func(*stscreds.WebIde
 	}
 
 	return func(opt *stscreds.WebIdentityRoleOptions) {}
-}
-
-// TODO(muvaf): All the types that use CreateJSONPatch are known during
-// development time. In order to avoid unnecessary panic checks, we can generate
-// the code that creates a patch between two objects that share the same type.
-
-// CreateJSONPatch creates a diff JSON object that can be applied to any other
-// JSON object.
-func CreateJSONPatch(source, destination interface{}) ([]byte, error) {
-	sourceJSON, err := json.Marshal(source)
-	if err != nil {
-		return nil, err
-	}
-	destinationJSON, err := json.Marshal(destination)
-	if err != nil {
-		return nil, err
-	}
-	patchJSON, err := jsonpatch.CreateMergePatch(sourceJSON, destinationJSON)
-	if err != nil {
-		return nil, err
-	}
-	return patchJSON, nil
-}
-
-// String converts the supplied string for use with the AWS Go SDK.
-func String(v string, o ...FieldOption) *string {
-	for _, fo := range o {
-		if fo == FieldRequired && v == "" {
-			return aws.String(v)
-		}
-	}
-
-	if v == "" {
-		return nil
-	}
-
-	return aws.String(v)
-}
-
-// StringValue converts the supplied string pointer to a string, returning the
-// empty string if the pointer is nil.
-// TODO(muvaf): is this really meaningful? why not implement it?
-func StringValue(v *string) string {
-	return aws.ToString(v)
-}
-
-// StringSliceToPtr converts the supplied string array to an array of string pointers.
-func StringSliceToPtr(slice []string) []*string {
-	if slice == nil {
-		return nil
-	}
-
-	res := make([]*string, len(slice))
-	for i, s := range slice {
-		res[i] = String(s)
-	}
-	return res
-}
-
-// StringPtrSliceToValue converts the supplied string pointer array to an array of strings.
-func StringPtrSliceToValue(slice []*string) []string {
-	if slice == nil {
-		return nil
-	}
-
-	res := make([]string, len(slice))
-	for i, s := range slice {
-		res[i] = StringValue(s)
-	}
-	return res
-}
-
-// BoolValue calls underlying aws ToBool
-func BoolValue(v *bool) bool {
-	return aws.ToBool(v)
-}
-
-// Int64Value converts the supplied int64 pointer to a int64, returning
-// 0 if the pointer is nil.
-func Int64Value(v *int64) int64 {
-	if v != nil {
-		return *v
-	}
-	return 0
-}
-
-// Int32Value converts the supplied int32 pointer to a int32, returning
-// 0 if the pointer is nil.
-func Int32Value(v *int32) int32 {
-	if v != nil {
-		return *v
-	}
-	return 0
-}
-
-// TimeToMetaTime converts a standard Go time.Time to a K8s metav1.Time.
-func TimeToMetaTime(t *time.Time) *metav1.Time {
-	if t == nil {
-		return nil
-	}
-	return &metav1.Time{Time: *t}
-}
-
-// LateInitializeStringPtr returns in if it's non-nil, otherwise returns from
-// which is the backup for the cases in is nil.
-func LateInitializeStringPtr(in *string, from *string) *string {
-	if in != nil {
-		return in
-	}
-	return from
-}
-
-// LateInitializeString returns `from` if `in` is empty and `from` is non-nil,
-// in other cases it returns `in`.
-func LateInitializeString(in string, from *string) string {
-	if in == "" && from != nil {
-		return *from
-	}
-	return in
-}
-
-// LateInitializeTimePtr returns in if it's non-nil, otherwise returns from
-// which is the backup for the cases in is nil.
-func LateInitializeTimePtr(in *metav1.Time, from *time.Time) *metav1.Time {
-	if in != nil {
-		return in
-	}
-	if from != nil {
-		t := metav1.NewTime(*from)
-		return &t
-	}
-	return nil
-}
-
-// Int64 converts the supplied int for use with the AWS Go SDK.
-func Int64(v int, o ...FieldOption) *int64 {
-	for _, fo := range o {
-		if fo == FieldRequired && v == 0 {
-			return aws.Int64(int64(v))
-		}
-	}
-
-	if v == 0 {
-		return nil
-	}
-
-	return aws.Int64(int64(v))
-}
-
-// Int32 converts the supplied int for use with the AWS Go SDK.
-func Int32(v int, o ...FieldOption) *int32 {
-	for _, fo := range o {
-		if fo == FieldRequired && v == 0 {
-			return aws.Int32(int32(v))
-		}
-	}
-
-	if v == 0 {
-		return nil
-	}
-
-	return aws.Int32(int32(v))
-}
-
-// Int64Address returns the given *int in the form of *int64.
-func Int64Address(i *int) *int64 {
-	if i == nil {
-		return nil
-	}
-	return aws.Int64(int64(*i))
-}
-
-// Int32Address returns the given *int in the form of *int32.
-func Int32Address(i *int) *int32 {
-	if i == nil {
-		return nil
-	}
-	return aws.Int32(int32(*i))
-}
-
-// IntAddress converts the supplied int64 pointer to an int pointer, returning nil if
-// the pointer is nil.
-func IntAddress(i *int64) *int {
-	if i == nil {
-		return nil
-	}
-	r := int(*i)
-	return &r
-}
-
-// IntFrom32Address converts the supplied int32 pointer to an int pointer, returning nil if
-// the pointer is nil.
-func IntFrom32Address(i *int32) *int {
-	if i == nil {
-		return nil
-	}
-	r := int(*i)
-	return &r
-}
-
-// LateInitializeIntPtr returns in if it's non-nil, otherwise returns from
-// which is the backup for the cases in is nil.
-func LateInitializeIntPtr(in *int, from *int64) *int {
-	if in != nil {
-		return in
-	}
-	if from != nil {
-		i := int(*from)
-		return &i
-	}
-	return nil
-}
-
-// LateInitializeIntFrom32Ptr returns in if it's non-nil, otherwise returns from
-// which is the backup for the cases in is nil.
-// This function considered that nil and 0 values are same. However, for a *int32, nil and 0 values must be different
-// because if the external AWS resource has a field with 0 value, during late initialization setting this value
-// in CR must be allowed. Please see the LateInitializeIntFromInt32Ptr func.
-func LateInitializeIntFrom32Ptr(in *int, from *int32) *int {
-	if in != nil {
-		return in
-	}
-	if from != nil && *from != 0 {
-		i := int(*from)
-		return &i
-	}
-	return nil
-}
-
-// LateInitializeIntFromInt32Ptr returns in if it's non-nil, otherwise returns from
-// which is the backup for the cases in is nil.
-func LateInitializeIntFromInt32Ptr(in *int, from *int32) *int {
-	if in != nil {
-		return in
-	}
-
-	if from != nil {
-		i := int(*from)
-		return &i
-	}
-
-	return nil
-}
-
-// LateInitializeInt32Ptr returns in if it's non-nil, otherwise returns from
-// which is the backup for the cases in is nil.
-func LateInitializeInt32Ptr(in *int32, from *int32) *int32 {
-	if in != nil {
-		return in
-	}
-	return from
-}
-
-// LateInitializeInt64Ptr returns in if it's non-nil, otherwise returns from
-// which is the backup for the cases in is nil.
-func LateInitializeInt64Ptr(in *int64, from *int64) *int64 {
-	if in != nil {
-		return in
-	}
-	return from
-}
-
-// LateInitializeInt32 returns in if it's non-zero, otherwise returns from
-// which is the backup for the cases in is zero.
-func LateInitializeInt32(in int32, from int32) int32 {
-	if in != 0 {
-		return in
-	}
-	return from
-}
-
-// LateInitializeInt64 returns in if it's non-zero, otherwise returns from
-// which is the backup for the cases in is zero.
-func LateInitializeInt64(in int64, from int64) int64 {
-	if in != 0 {
-		return in
-	}
-	return from
-}
-
-// LateInitializeStringPtrSlice returns in if it's non-nil or from is zero
-// length, otherwise it returns from.
-func LateInitializeStringPtrSlice(in []*string, from []*string) []*string {
-	if in != nil || len(from) == 0 {
-		return in
-	}
-
-	return from
-}
-
-// LateInitializeInt64PtrSlice returns in if it's non-nil or from is zero
-// length, otherwise it returns from.
-func LateInitializeInt64PtrSlice(in []*int64, from []*int64) []*int64 {
-	if in != nil || len(from) == 0 {
-		return in
-	}
-
-	return from
-}
-
-// Bool converts the supplied bool for use with the AWS Go SDK.
-func Bool(v bool, o ...FieldOption) *bool {
-	for _, fo := range o {
-		if fo == FieldRequired && !v {
-			return aws.Bool(v)
-		}
-	}
-
-	if !v {
-		return nil
-	}
-	return aws.Bool(v)
-}
-
-// LateInitializeBoolPtr returns in if it's non-nil, otherwise returns from
-// which is the backup for the cases in is nil.
-func LateInitializeBoolPtr(in *bool, from *bool) *bool {
-	if in != nil {
-		return in
-	}
-	return from
-}
-
-// CompactAndEscapeJSON removes space characters and URL-encodes the JSON string.
-func CompactAndEscapeJSON(s string) (string, error) {
-	buffer := new(bytes.Buffer)
-	if err := json.Compact(buffer, []byte(s)); err != nil {
-		return "", err
-	}
-	return url.QueryEscape(buffer.String()), nil
-}
-
-// DiffTags returns tags that should be added or removed.
-func DiffTags(local, remote map[string]string) (add map[string]string, remove []string) {
-	add = make(map[string]string, len(local))
-	remove = []string{}
-	for k, v := range local {
-		add[k] = v
-	}
-	for k, v := range remote {
-		switch val, ok := local[k]; {
-		case ok && val != v:
-			remove = append(remove, k)
-		case !ok:
-			remove = append(remove, k)
-			delete(add, k)
-		default:
-			delete(add, k)
-		}
-	}
-	return
-}
-
-// DiffEC2Tags returns []ec2type.Tag that should be added or removed.
-func DiffEC2Tags(local []ec2type.Tag, remote []ec2type.Tag) (add []ec2type.Tag, remove []ec2type.Tag) {
-	var tagsToAdd = make(map[string]string, len(local))
-	add = []ec2type.Tag{}
-	remove = []ec2type.Tag{}
-	for _, j := range local {
-		tagsToAdd[aws.ToString(j.Key)] = aws.ToString(j.Value)
-	}
-	for _, j := range remote {
-		switch val, ok := tagsToAdd[aws.ToString(j.Key)]; {
-		case ok && val == aws.ToString(j.Value):
-			delete(tagsToAdd, aws.ToString(j.Key))
-		case !ok:
-			remove = append(remove, ec2type.Tag{
-				Key:   j.Key,
-				Value: nil,
-			})
-		}
-	}
-	for i, j := range tagsToAdd {
-		add = append(add, ec2type.Tag{
-			Key:   aws.String(i),
-			Value: aws.String(j),
-		})
-	}
-	return
-}
-
-// DiffLabels returns labels that should be added, modified, or removed.
-func DiffLabels(local, remote map[string]string) (addOrModify map[string]string, remove []string) {
-	addOrModify = make(map[string]string, len(local))
-	remove = []string{}
-	for k, v := range local {
-		addOrModify[k] = v
-	}
-	for k, v := range remote {
-		switch val, ok := local[k]; {
-		case ok && val != v:
-			// if value does not match key it will be updated by the correct
-			// key-value pair being present in the returned addOrModify map
-			continue
-		case !ok:
-			remove = append(remove, k)
-			delete(addOrModify, k)
-		default:
-			delete(addOrModify, k)
-		}
-	}
-	return
-}
-
-// IsPolicyUpToDate Marshall policies to json for a compare to get around string ordering
-func IsPolicyUpToDate(local, remote *string) bool {
-	var localUnmarshalled interface{}
-	var remoteUnmarshalled interface{}
-
-	var err error
-	err = json.Unmarshal([]byte(StringValue(local)), &localUnmarshalled)
-	if err != nil {
-		return false
-	}
-	err = json.Unmarshal([]byte(StringValue(remote)), &remoteUnmarshalled)
-	if err != nil {
-		return false
-	}
-
-	sortSlicesOpt := cmpopts.SortSlices(func(x, y interface{}) bool {
-		if a, ok := x.(string); ok {
-			if b, ok := y.(string); ok {
-				return a < b
-			}
-		}
-		// Note: Unknown types in slices will not cause a panic, but
-		// may not be sorted correctly. Depending on how AWS handles
-		// these, it may cause constant updates - but better this than
-		// panicing.
-		return false
-	})
-	return cmp.Equal(localUnmarshalled, remoteUnmarshalled, cmpopts.EquateEmpty(), sortSlicesOpt)
-}
-
-// Wrap will remove the request-specific information from the error and only then
-// wrap it.
-func Wrap(err error, msg string) error {
-	// NOTE(muvaf): nil check is done for performance, otherwise errors.As makes
-	// a few reflection calls before returning false, letting awsErr be nil.
-	if err == nil {
-		return nil
-	}
-	var awsErr smithy.APIError
-	if errors.As(err, &awsErr) {
-		return errors.Wrap(awsErr, msg)
-	}
-	// AWS SDK v1 uses different interfaces than v2 and it doesn't unwrap to
-	// the underlying error. So, we need to strip off the unique request ID
-	// manually.
-	if v1RequestError, ok := err.(awserr.RequestFailure); ok { //nolint:errorlint
-		// TODO(negz): This loses context about the underlying error
-		// type, preventing us from using errors.As to figure out what
-		// kind of error it is. Could we do this without losing
-		// context?
-		return errors.Wrap(errors.New(strings.ReplaceAll(err.Error(), v1RequestError.RequestID(), "")), msg)
-	}
-	return errors.Wrap(err, msg)
-}
-
-// DiffTagsMapPtr returns which AWS Tags exist in the resource tags and which are outdated and should be removed
-func DiffTagsMapPtr(spec map[string]*string, current map[string]*string) (map[string]*string, []*string) {
-	addMap := make(map[string]*string, len(spec))
-	removeTags := make([]*string, 0)
-	for k, v := range current {
-		if StringValue(spec[k]) == StringValue(v) {
-			continue
-		}
-		removeTags = append(removeTags, String(k))
-	}
-	for k, v := range spec {
-		if StringValue(current[k]) == StringValue(v) {
-			continue
-		}
-		addMap[k] = v
-	}
-	return addMap, removeTags
-}
-
-// CIDRBlocksEqual returns whether or not two CIDR blocks are equal:
-// - Both CIDR blocks parse to an IP address and network
-// - The string representation of the IP addresses are equal
-// - The string representation of the networks are equal
-func CIDRBlocksEqual(cidr1, cidr2 string) bool {
-	ip1, ipnet1, err := net.ParseCIDR(cidr1)
-	if err != nil {
-		return false
-	}
-	ip2, ipnet2, err := net.ParseCIDR(cidr2)
-	if err != nil {
-		return false
-	}
-
-	return ip2.String() == ip1.String() && ipnet2.String() == ipnet1.String()
 }
